+++ query/mod.ts
@@ -2,5 +2,5 @@ export * from "./pipe.ts";
 export * from "./api.ts";
 export * from "./types.ts";
 export * from "./fetch.ts";
-export * as mdw from "./middleware.ts";
+export * as mdw from "./mdw.ts";
 export * from "./create-key.ts";
